fixed warning on return of size_t -1

+++ ref-impl/src/OM/utf8.cpp
@@ -163,6 +163,7 @@ int u8swcslen( const char* pu )
 // follows signature of size_t mbstowcs( wchar_t *, const char *, size_t count )
 size_t u8stowcs( wchar_t *pw, const char *pu, size_t count )
 {
+
 	size_t clen=0;
 
 	if( NULL==pw ) return u8swcslen( pu );
@@ -170,7 +171,7 @@ size_t u8stowcs( wchar_t *pw, const char *pu, size_t count )
 	while( *pu && clen<count )
 	{
 		int ulen=u8towc( &pw[clen], pu, 1 );
-		if( ulen<0 ) return -1;
+		if( ulen<0 ) return (size_t) -1;
 		else { clen++; pu+=ulen; }
 	}
 	if( '\0'==*pu && clen<count) pw[clen++]=L'\0';
